simple math trick
we see this picture and we see 10 number have in row number 1 and column number is 4.
if we find this number sqrt and use ceil function so now we find sqrt is 4. so we find our column = sqrt. and now my work is how to way we find this row. so if we use row = number - (sqrt -1) * (sqrt - 1)  = 10 - (4 - 1) * (4 - 1) = 10 - 3*3 = 10 - 9 = 1. so we find out expected output is row = 1 and our column = 4.

if our input is 25 so now we find our row and column in this number.
our extra number = sqrt(25)* sqrt(25) - 25 = 5*5 - 25 = 0.  so our column = 5 and row = extra number + 1 = 0+1 = 1.